Combine ui-obj and ui-object, tidy
