+++ src/ui-death.c
@@ -30,7 +30,7 @@
 #include "ui-input.h"
 #include "ui-knowledge.h"
 #include "ui-menu.h"
-#include "ui-obj.h"
+#include "ui-object.h"
 #include "ui-player.h"
 #include "ui-score.h"
 #include "wizard.h"
