+++ src/ui-game.c
@@ -38,7 +38,6 @@
 #include "ui-keymap.h"
 #include "ui-knowledge.h"
 #include "ui-map.h"
-#include "ui-obj.h"
 #include "ui-object.h"
 #include "ui-output.h"
 #include "ui-player.h"
