+++ src/ui-input.c
@@ -34,7 +34,6 @@
 #include "ui-keymap.h"
 #include "ui-knowledge.h"
 #include "ui-map.h"
-#include "ui-obj.h"
 #include "ui-object.h"
 #include "ui-output.h"
 #include "ui-player.h"
