+++ src/ui-knowledge.c
@@ -43,7 +43,7 @@
 #include "ui-menu.h"
 #include "ui-mon-list.h"
 #include "ui-mon-lore.h"
-#include "ui-obj.h"
+#include "ui-object.h"
 #include "ui-obj-list.h"
 #include "ui-options.h"
 #include "ui-output.h"
