+++ src/ui-map.c
@@ -27,7 +27,7 @@
 #include "trap.h"
 #include "ui-input.h"
 #include "ui-map.h"
-#include "ui-obj.h"
+#include "ui-object.h"
 #include "ui-output.h"
 #include "ui-prefs.h"
 #include "ui-term.h"
