+++ src/ui-player.c
@@ -32,7 +32,7 @@
 #include "ui-history.h"
 #include "ui-input.h"
 #include "ui-menu.h"
-#include "ui-obj.h"
+#include "ui-object.h"
 #include "ui-output.h"
 #include "ui-player.h"
