+++ src/ui-target.c
@@ -35,7 +35,7 @@
 #include "ui-keymap.h"
 #include "ui-map.h"
 #include "ui-mon-lore.h"
-#include "ui-obj.h"
+#include "ui-object.h"
 #include "ui-output.h"
 #include "ui-target.h"
 #include "ui-term.h"
